Bump to 3.0.1 for release
I forgot to include this in https://github.com/alphagov/notifications-php-client/pull/80/files
so am including it here.

Wasn't sure if this was a major breaking change rather than a bug fix.
Technically no outward behaviour is difference which leads me to a patch
but also people can no longer use php-jwt version < 5.0.0 so could be
argued to be a major.

I've gone minor as for most people I'm sure this won't make any
difference in the slightest