Update docs to catch ApiException rather than NotifyException
As pointed out in
https://github.com/alphagov/notifications-php-client/issues/25,
the `NotifyException` class does not have a `getErrors` or `getCode`
method. By our docs currently saying that users should be default catch
the `NotifyException` class, then if a `NotifyException` that is not an
`ApiException` is thrown and the user tries to call `getErrors` or
`getCode` then another exception will be thrown.

This raises two possible solutions to fix
1) Add `getErrors` and `getCode` methods to `NotifyException. This
doesn't seem that appropriate though as not all types of
NotifyExceptions do have `errors` or an http error `code`.
2) Suggest users catch `ApiException`s when making API calls. This will
mean that a user will reliably be able to access `getErrors` and
`getCode`. The downside to this is that if another type of
`NotifyException` is thrown, then it will not be caught. However, this
would be no worse than our current situation and will only be uncaught
for errors that we really aren't expecting that should ideally be
investigated.

In the end, I went for option 2 as this seemed the least worst and also
was fairly quick to implement.