Change CONTRIBUTING.md to advocate using Docker
This can be consistent across the other API clients, some of which
are harder to run natively. Although the task to prepare the image
is currently run automatically, this will change in future commits,
so we may as well put in the words for it now.

It's unclear why we didn't refer to tasks in the Makefile instead
of writing out the full commands. Using "make test..." is consistent
with our other API clients.