DRY-up running Docker commands with a script
This matches the approach we take in other repos, and has the added
benefit of being able to easily run ad-hoc commands, without needing
to make new tasks for them in the Makefile e.g. individual tests.

Note that I've also removed the task to cleanup Docker containers,
which should be unnecessary due to the use of the '--rm' flag.