Remove documentation for cURL support
Support for cURL was removed in version 3.0.0 of this client
without us realising.

To make the docs correct, we are therefore removing the instructions
on how to use it.

We may decide to add back in cURL support but haven't commited to
doing so right at this point.

For more context, see:
https://github.com/alphagov/notifications-php-client/issues/111#issuecomment-1165767985