Refactor sendEmail integration test to reduce duplication
We had 2 identical tests. One that sends an email with a reply to
email address set, and the other that didn't. The assertions for
the 2 tests were completely the same - it wasn't really checking
that the API would be doing any different behaviour based on
the input so I didn't feel the 2 tests were particularly beneficial.

I suggest rather than continuing the pattern of one single integration
test for each optional parameter (there wasn't one for reference
but there was one for reply to address), we instead have a single
integration test that we supply all the optional parameters to
and check comes back with responses based on those parameters.

One benefit of this is that integration tests are slow, so this
should slightly reduce the time for them to run, rather than keep
adding slow tests for every parameter we add.

.